Implement "packs" architecture.
We have a problem. Crates.io doesn't let us upload crates bigger than 10MiB, but
a single language.c can be over 20MiB. So, not only can we not upload all of the
compiled language.c files, but we can't upload any of them!

This commit implements a strategy by which the languages are split up into
"pack" crates and then tarred and compressed using zstd. The pack crates are
generated automatically, as are the packs and cross-dependencies. Each pack is
sub 10MiB, allowing for publishing on crates.io. Each pack's build script
decompressed, untars, and compiles the relevant files, making the compiled
artifacts available as exports.

The "real" crate, `jellybean`, imports all of the languages from the packs
transparently and makes them avaialble with a nicer API.

In all, this makes jellybean the first truly usable alternative to syntect.